requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
jinja2
python-dotenv
psycopg2-binary
